wip: refactoring - remove 'utils' folder - move code to utils.lua
